+++ Buildenator/Configuration/EntityToBuild.cs
@@ -5,6 +5,8 @@ using Microsoft.CodeAnalysis;
 using System.Collections.Generic;
 using System.Linq;
 using Buildenator.Abstraction;
+using System.Diagnostics;
+using System.Reflection.Metadata;
 
 namespace Buildenator.Configuration;
 
@@ -13,7 +15,7 @@ internal sealed class EntityToBuild : IEntityToBuild
     public string Name { get; }
     public string FullName { get; }
     public string FullNameWithConstraints { get; }
-    public Constructor? ConstructorToBuild { get; }
+    public Constructor ConstructorToBuild { get; }
     public IReadOnlyList<TypedSymbol> SettableProperties { get; }
     public IReadOnlyList<TypedSymbol> ReadOnlyProperties { get; }
     public string[] AdditionalNamespaces { get; }
@@ -50,11 +52,6 @@ internal sealed class EntityToBuild : IEntityToBuild
 
     public IReadOnlyList<ITypedSymbol> GetAllUniqueSettablePropertiesAndParameters()
     {
-        if (ConstructorToBuild is null)
-        {
-            return _uniqueTypedSymbols ??= SettableProperties;
-        }
-
         return _uniqueTypedSymbols ??= SettableProperties
             .Where(x => !ConstructorToBuild.ContainsParameter(x.SymbolName))
             .Concat(ConstructorToBuild.Parameters).ToList();
@@ -62,11 +59,6 @@ internal sealed class EntityToBuild : IEntityToBuild
 
     public IReadOnlyList<ITypedSymbol> GetAllUniqueReadOnlyPropertiesWithoutConstructorsParametersMatch()
     {
-        if (ConstructorToBuild is null)
-        {
-            return _uniqueReadOnlyTypedSymbols ??= ReadOnlyProperties;
-        }
-
         return _uniqueReadOnlyTypedSymbols ??= ReadOnlyProperties
             .Where(x => !ConstructorToBuild.ContainsParameter(x.SymbolName)).ToList();
     }
@@ -86,28 +78,38 @@ internal sealed class EntityToBuild : IEntityToBuild
 
     internal sealed class Constructor
     {
-        public static Constructor? CreateConstructorOrDefault(
+        public bool IsPrivate { get; }
+
+        public static Constructor CreateConstructorOrDefault(
             INamedTypeSymbol entityToBuildSymbol,
             IMockingProperties? mockingConfiguration,
             IFixtureProperties? fixtureConfiguration,
             NullableStrategy nullableStrategy)
         {
-            var onlyPublicOrInternalConstructors = entityToBuildSymbol.Constructors
-                .Where(m => 
-                    !m.IsStatic
-                    && !m.IsImplicitlyDeclared
-                    && (m.DeclaredAccessibility == Accessibility.Public || m.DeclaredAccessibility == Accessibility.Internal))
+            var constructors = entityToBuildSymbol.Constructors.Select(a => a).ToArray();
+            var onlyPublicConstructors = constructors
+                .Where(m => m.DeclaredAccessibility == Accessibility.Public || m.DeclaredAccessibility == Accessibility.Internal)
                 .ToList();
 
-            return onlyPublicOrInternalConstructors.Count > 0
-                    ? new Constructor(onlyPublicOrInternalConstructors.OrderByDescending(x => x.Parameters.Length).First().Parameters
-                        .ToDictionary(x => x.PascalCaseName(), s => new TypedSymbol(s, mockingConfiguration, fixtureConfiguration, nullableStrategy)))
-                    : default;
+            var isPrivate = onlyPublicConstructors.Count == 0;
+
+            Dictionary<string, TypedSymbol> parameters = [];
+            if (!isPrivate)
+            {
+                parameters = onlyPublicConstructors
+                    .OrderByDescending(x => x.Parameters.Length)
+                    .First()
+                    .Parameters
+                    .ToDictionary(x => x.PascalCaseName(), s => new TypedSymbol(s, mockingConfiguration, fixtureConfiguration, nullableStrategy));
+            }
+
+            return new Constructor(parameters, isPrivate);
         }
 
-        private Constructor(IReadOnlyDictionary<string, TypedSymbol> constructorParameters)
+        private Constructor(IReadOnlyDictionary<string, TypedSymbol> constructorParameters, bool isPrivate)
         {
             ConstructorParameters = constructorParameters;
+            IsPrivate = isPrivate;
         }
 
         public IReadOnlyDictionary<string, TypedSymbol> ConstructorParameters { get; }