+++ Tests/Buildenator.IntegrationTests.SharedEntities/EntityWithImplicitConstructor.cs
@@ -4,5 +4,5 @@ public class EntityWithImplicitConstructor
 {
 
     public int PropertyIntGetter { get; } = 1;
-    public string PropertyGetter { get; } = "1";
+    public string PropertyGetter { get; set; }
 }